Moved route definitions into env wrapped check, webpackified embed
